comments

+++ zoo/src/main/scala/com/intel/analytics/zoo/pipeline/api/net/python/PythonZooNet.scala
@@ -160,8 +160,10 @@ class PythonZooNet[T: ClassTag](implicit ev: TensorNumeric[T]) extends PythonZoo
           override def run(): Unit = {
             // Give it a chance to be gracefully killed
             killPids(processToBeKill, "kill ")
-            Thread.sleep(2000)
-            killPids(processToBeKill, "kill -9")
+            if (processToBeKill.isEmpty) {
+              Thread.sleep(2000)
+              killPids(processToBeKill, "kill -9")
+            }
           }
       })
   }
